+++ homeassistant/components/lcn/manifest.json
@@ -3,8 +3,10 @@
   "name": "Lcn",
   "documentation": "https://www.home-assistant.io/components/lcn",
   "requirements": [
-    "pypck==0.5.9"
+    "pypck==0.6.0"
   ],
   "dependencies": [],
-  "codeowners": []
+  "codeowners": [
+    "@alengwenus"
+  ]
 }
